+++ opencolorio_config_aces/clf/__init__.py
@@ -5,10 +5,10 @@ from .discover import (discover_clf_transforms, classify_clf_transforms,
                        unclassify_clf_transforms, filter_clf_transforms,
                        print_clf_taxonomy)
 
-from .transforms import matrix_3x3_to_4x4, generate_clf
+from .transforms import generate_clf
 
 __all__ = [
     'discover_clf_transforms', 'classify_clf_transforms',
     'unclassify_clf_transforms', 'filter_clf_transforms', 'print_clf_taxonomy'
 ]
-__all__ += ['matrix_3x3_to_4x4', 'generate_clf']
+__all__ += ['generate_clf']
