+++ opencolorio_config_aces/clf/transforms/__init__.py
@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright Contributors to the OpenColorIO Project.
 
-from .common import matrix_3x3_to_4x4, generate_clf
+from .common import generate_clf
 
-__all__ = ['matrix_3x3_to_4x4', 'generate_clf']
+__all__ = ['generate_clf']
