+++ opencolorio_config_aces/utilities/__init__.py
@@ -1,16 +1,17 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright Contributors to the OpenColorIO Project.
 
-from .common import (
-    DocstringDict, first_item, common_ancestor, paths_common_ancestor,
-    vivification, vivified_to_dict, message_box, is_colour_installed,
-    is_jsonpickle_installed, is_networkx_installed, is_opencolorio_installed,
-    REQUIREMENTS_TO_CALLABLE, required, is_string, is_iterable, git_describe)
+from .common import (DocstringDict, first_item, common_ancestor,
+                     paths_common_ancestor, vivification, vivified_to_dict,
+                     message_box, is_colour_installed, is_jsonpickle_installed,
+                     is_networkx_installed, is_opencolorio_installed,
+                     REQUIREMENTS_TO_CALLABLE, required, is_string,
+                     is_iterable, git_describe, matrix_3x3_to_4x4)
 
 __all__ = [
     'DocstringDict', 'first_item', 'common_ancestor', 'paths_common_ancestor',
     'vivification', 'vivified_to_dict', 'message_box', 'is_colour_installed',
     'is_jsonpickle_installed', 'is_networkx_installed',
     'is_opencolorio_installed', 'REQUIREMENTS_TO_CALLABLE', 'required',
-    'is_string', 'is_iterable', 'git_describe'
+    'is_string', 'is_iterable', 'git_describe', 'matrix_3x3_to_4x4'
 ]
